requirements.txt
# Core Dependencies
bedrock-agentcore==0.1.7
boto3>=1.34.0
botocore>=1.34.0

# UI
streamlit>=1.35.0

# Data Processing
pydantic>=2.7.0
python-dateutil>=2.9.0
pytz>=2024.1

# Utilities
python-dotenv>=1.0.0
requests>=2.32.0
pyyaml>=6.0

